factor out campaign projection
