+++ shared/banner_functions.js
@@ -70,58 +70,6 @@ function getDaysRemaining( language ) {
 	return daysRemaining + ' ' + ( daysRemaining > 1 ? messages[ lang ].days : messages[ lang ].day );
 }
 
-function getSecsPassed() {
-	var startDate = baseDate.split( '-' ),
-		startDateObj = new Date( startDate[ 0 ], startDate[ 1 ] - 1, startDate[ 2 ] ),
-		maxSecs = Math.floor( new Date( finalDateTime - startDateObj ) / 1000 ),
-		secsPassed = Math.floor( ( new Date() - startDateObj ) / 1000 );
-
-	if ( secsPassed < 0 ) {
-		secsPassed = 0;
-	}
-	if ( secsPassed > maxSecs ) {
-		secsPassed = maxSecs;
-	}
-
-	return secsPassed;
-}
-
-function getApprDonationsRaw( rand ) {
-	var startDonations = collectedBase,
-		secsPast = getSecsPassed();
-
-	return startDonations + getApprDonationsFor( secsPast, rand );
-}
-
-function getApprDonatorsRaw( rand ) {
-	var startDonators = donorsBase,
-		secsPast = getSecsPassed();
-
-	return startDonators + getApprDonatorsFor( secsPast, rand );
-}
-
-function getApprDonationsFor( secsPast, rand ) {
-	var apprDontionsMinute = donationsPerMinApproximation,
-		randFactor = 0;
-
-	if ( rand === true ) {
-		randFactor = Math.floor( ( Math.random() ) + 0.5 - 0.2 );
-	}
-
-	return ( secsPast / 60 * ( apprDontionsMinute * ( 100 + randFactor ) ) / 100 );
-}
-
-function getApprDonatorsFor( secsPast, rand ) {
-	var apprDonatorsMinute = donorsPerMinApproximation,
-		randFactor = 0;
-
-	if ( rand === true ) {
-		randFactor = Math.floor( ( Math.random() ) + 0.5 - 0.2 );
-	}
-
-	return ( secsPast / 60 * ( apprDonatorsMinute * ( 100 + randFactor ) ) / 100 );
-}
-
 function getCurrentGermanDay() {
 	switch ( new Date().getDay() ) {
 		case 0:
@@ -173,18 +121,6 @@ function getDigitGroupingCharacter() {
 		default:
 			return '.';
 	}
-}
-
-function addPointsToNum( num ) {
-	// jscs:disable disallowImplicitTypeConversion
-	num = parseInt( num, 10 ) + '';
-	// jscs:enable disallowImplicitTypeConversion
-	num = num.replace( /\./g, ',' );
-	return num.replace( /(\d)(?=(\d\d\d)+(?!\d))/g, '$1' + getDigitGroupingCharacter() );
-}
-
-function floorF( num ) {
-	return Math.floor( num * 100 ) / 100;
 }
 
 function getImpCount() {
